+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.4.0] - 2020-08-07
+
 ### Added
 
 - Builder support for api filters and filter properties
@@ -169,7 +171,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - base app
 
-[Unreleased]: https://github.com/christiansiewert/aaas-api/compare/v1.3.0...develop
+[Unreleased]: https://github.com/christiansiewert/aaas-api/compare/v1.4.0...develop
+[1.4.0]: https://github.com/christiansiewert/aaas-api/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/christiansiewert/aaas-api/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/christiansiewert/aaas-api/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/christiansiewert/aaas-api/compare/v1.0.0...v1.1.0